Part3c working FINALLY
