+++ .gitignore
@@ -28,3 +28,6 @@
 /A1P3c-Test-Threading/build/
 /A1P3c-JustServer/nbproject/private/
 /A1P3c-JustServer/build/
+/A1P3c-Client1/nbproject/private/
+/A1P3c-Client1/build/
+/A1P3c-orig/nbproject/private/
