--- A1P3c-JustServer/src/ServerPkg/names.txt
@@ -1,15 +1,0 @@
-Dan
-Batman
-The Incredible Hulk
-Spiderman
-Mary Jane
-Obi-Kenobi
-Wolverine
-Gene Grey
-The Professor
-SliderGambit
-Bob M.
-John Butler
-Eddie Vedder
-Chris Cornell
-Bart Simpson
