+++ A1P3c-Test-Threading/src/ThreadTest/ProcessThreadTest.java
@@ -7,7 +7,8 @@ import java.net.*;
 
 public class ProcessThreadTest implements Runnable {
     // variables
-    Socket sock;
+    final Socket sock;
+    int clientNum;
 
     // constructor
     public ProcessThreadTest (Socket s) {
@@ -16,20 +17,17 @@ public class ProcessThreadTest implements Runnable {
     
     @Override
     public void run() {
-        String inString;
-        ObjectOutputStream outStream = null;
+        String inString = "";
         
         try {
             // initiate streams
-            outStream = new ObjectOutputStream(sock.getOutputStream());
+            ObjectOutputStream outStream = new ObjectOutputStream(sock.getOutputStream());
             outStream.flush();
             BufferedReader inStream =
                     new BufferedReader(
                             new InputStreamReader(sock.getInputStream()));
-            System.out.println("Initiated streams");
             
             // client interaction
-            inString = inStream.toString(); // read input from client
             while (!inString.equals("5")) { // loop until exit option
                 inString = inStream.readLine(); // read menu option from client
                 int menuOption = Integer.parseInt(inString); // set menu option
@@ -37,26 +35,24 @@ public class ProcessThreadTest implements Runnable {
                 // if not exit option, process client request
                 if (!inString.equals("5")) {
                     
-                    // prompt client for name
-                    outStream.writeObject(promptClient(menuOption));
-                    outStream.flush();
+                    synchronized(namesList) {
+                        // prompt client for name
+                        outStream.writeObject(promptClient(menuOption));
+                        outStream.flush();
 
-                    inString = inStream.readLine(); // read input from client
+                        inString = inStream.readLine(); // read input from client
 
-                    // send processed result to client
-                    outStream.writeObject(menuCommand(menuOption, inString));
-                    outStream.flush();
+                        // send processed result to client
+                        outStream.writeObject(menuCommand(menuOption, inString));
+                        outStream.flush();
+                        saveToFile();
+                    }
                 }
             }
             // close connection to client
             System.out.println("Closed connection for "
                 +sock.getInetAddress().getHostAddress());
-            inString = ""; // reset for next connection
             sock.close(); // close socket
-
-            // update names file
-            saveToFile();
-            System.out.println("Updated/saved names file.");
         } 
         catch (IOException ioe) {
             System.out.println("** Error handling client interaction.");
@@ -69,7 +65,7 @@ public class ProcessThreadTest implements Runnable {
                 System.out.println("** Error saving names file.");
             }
         }
-    }
+    } // end run
 
     // initiates menu command request
     public static String menuCommand(int i, String cInput) {
