+++ A1P3c-Test-Threading/src/ThreadTest/names.txt
@@ -13,3 +13,4 @@ John Butler
 Eddie Vedder
 Chris Cornell
 Bart Simpson
+New name
