+++ A1P3c/src/Part3c/ProcessClient.java
@@ -1,19 +1,22 @@
-//System.out.println("");
+/*
+    Dan Hogan
+    ISY00246
+    Assignment 1
+    Part 3c
+    Purpose: Server class that allows multithreading processing with names file
+ */
 package Part3c;
 
+import static Part3c.ThreadServer.saveToFile;
 import static Part3c.ThreadServer.namesList;
 import java.io.*;
 import java.net.*;
 
-/**
- *
- * @author Ðan
- */
-public class ProcessClient implements Runnable{
+public class ProcessClient implements Runnable {
     // variables
-    String inString = "";
-    Socket sock;
-    
+    final Socket sock;
+    int clientNum;
+
     // constructor
     public ProcessClient (Socket s) {
         this.sock = s;
@@ -21,63 +24,56 @@ public class ProcessClient implements Runnable{
     
     @Override
     public void run() {
-        // create in and out streams
+        String inString = "";
+        
         try {
-            ObjectOutputStream outStream = null;
-            outStream = 
-                new ObjectOutputStream(sock.getOutputStream());
+            // initiate streams
+            ObjectOutputStream outStream = 
+                    new ObjectOutputStream(sock.getOutputStream());
             outStream.flush();
             BufferedReader inStream =
                     new BufferedReader(
                             new InputStreamReader(sock.getInputStream()));
-            System.out.println("Created streams");
-
-            System.out.println("Entering 'while !inString equals 5' loop");
+            
+            // client interaction
             while (!inString.equals("5")) { // loop until exit option
                 inString = inStream.readLine(); // read menu option from client
                 int menuOption = Integer.parseInt(inString); // set menu option
 
-                System.out.println("Entering 'if inString = 5'");
+                // if not exit option, process client request
                 if (!inString.equals("5")) {
-                    // prompt client for name
-                    outStream.writeObject(promptClient(menuOption));
-                    outStream.flush();
+                    
+                    synchronized(namesList) {
+                        // prompt client for name
+                        outStream.writeObject(promptClient(menuOption));
+                        outStream.flush();
 
-                    inString = inStream.readLine(); // read input from client
+                        inString = inStream.readLine(); // read input from client
 
-                    // send result to client
-                    outStream.writeObject(menuCommand(menuOption, inString));
-                    outStream.flush();
+                        // send processed result to client
+                        outStream.writeObject(menuCommand(menuOption, inString));
+                        outStream.flush();
+                        saveToFile();
+                    }
                 }
             }
-
-                            // close connection to client
+            // close connection to client
             System.out.println("Closed connection for "
                 +sock.getInetAddress().getHostAddress());
-            inString = ""; // reset for next connection
             sock.close(); // close socket
-
-            // update names file
-            saveToFile();
-            System.out.println("Updated/saved names file.");
-        } // end try
+        } 
         catch (IOException ioe) {
-            System.out.println("Error handling client interaction.");
-            ioe.printStackTrace();
+            System.out.println("** Error handling client interaction.");
         }
-    }
-        
-    // saves the list to a file (overwrites previous)
-    public static void saveToFile() throws IOException {
-        File file = new File("src/names.txt");
-        try (FileWriter fw = new FileWriter(file, false)) {
-            for (String s : namesList) {
-                if (!s.equals("")) {
-                    fw.write(s+"\n");
-                }
-            } // end for
-        } // end try
-    } // end save to file()
+        finally {
+            try {
+                saveToFile();
+                System.out.println("Updated/saved names file.");
+            } catch (IOException ex) {
+                System.out.println("** Error saving names file.");
+            }
+        }
+    } // end run
 
     // initiates menu command request
     public static String menuCommand(int i, String cInput) {
@@ -155,4 +151,4 @@ public class ProcessClient implements Runnable{
         return prompt;
     } // end promptClient()
     
-} // end class
+}
